# feat: improved performance of LIKE matching